% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_risk.R
\name{calculate_risk}
\alias{calculate_risk}
\title{Calculate risk measures}
\usage{
calculate_risk(
  psa_data,
  tmin,
  tmax,
  Dt,
  Dt_max = TRUE,
  W = FALSE,
  weight = NULL
)
}
\arguments{
\item{psa_data}{A data.frame (or a list of data.frames containing) where the first column contains the model time, and subsequent columns contain the predicted output for each simulation run at the respective time. See Data format section for more information.}

\item{tmin}{Numeric value giving the minimum simulation time}

\item{tmax}{Numeric value giving the maximum simulation time}

\item{Dt}{Numeric vector of decision thresholds at each time between \code{tmin} and `tmax``}

\item{Dt_max}{Logical indicating whether decision threshold is a maximum (\code{TRUE}) or minimum (\code{FALSE})}

\item{W}{Logical indicating whether the risk calculation should be weighted or not}

\item{weight}{Numeric vector containing weights for the risk calculation, one per time step}
}
\value{
Numeric value or list of risk score(s)
}
\description{
Calculate risk measures
}
\section{Data format}{


The model time in the first column must contain numeric values indicating a simulation time (ex. 0, 1, 2, 3,...) or dates (ex. 01/01/2024, 01/02/2024) which must be in \code{R} Date format (i.e., class="Date"). If comparing simulation results across multiple scenarios, provide a list of data.frames, where each data.frame gives the output for one scenario.
}

