% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_risk.R
\name{calculate_risk_1}
\alias{calculate_risk_1}
\title{Calculate risk measure for a single set of simulations}
\usage{
calculate_risk_1(psa_data, tmin, tmax, Dt, Dt_max, W, weight)
}
\arguments{
\item{psa_data}{A data.frame (or a list of data.frames containing) where the first column contains the model time, and subsequent columns contain the predicted output for each simulation run at the respective time. See Data format section for more information.}

\item{tmin}{Numeric value giving the minimum simulation time}

\item{tmax}{Numeric value giving the maximum simulation time}

\item{Dt}{Numeric vector of decision thresholds at each time between \code{tmin} and `tmax``}

\item{Dt_max}{Logical indicating whether decision threshold is a maximum (\code{TRUE}) or minimum (\code{FALSE})}

\item{W}{Logical indicating whether the risk calculation should be weighted or not}

\item{weight}{Numeric vector containing weights for the risk calculation, one per time step}
}
\value{
Numeric value of risk score
}
\description{
Calculate risk measure for a single set of simulations
}
