% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rr.R
\name{calculate_rr}
\alias{calculate_rr}
\title{Calculate relative risk}
\usage{
calculate_rr(exp_risk_list)
}
\arguments{
\item{exp_risk_list}{Numeric vector of expected risk, where the first element corresponds to the baseline scenario}
}
\value{
Numeric vector of relative risk values for each scenario except the baseline
}
\description{
Calculate relative risk
}
