% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_time.R
\name{calculate_time_1}
\alias{calculate_time_1}
\title{Calculate time of threshold exceedance a single set of simulations}
\usage{
calculate_time_1(psa_data, tmin, tmax, Dt, Dt_max = TRUE)
}
\arguments{
\item{psa_data}{A data.frame (or a list of data.frames containing) where the first column contains the model time, and subsequent columns contain the predicted output for each simulation run at the respective time. See Data format section for more information.}

\item{tmin}{Numeric value giving the minimum simulation time}

\item{tmax}{Numeric value giving the maximum simulation time}

\item{Dt}{Numeric vector of decision thresholds at each time between \code{tmin} and `tmax``}

\item{Dt_max}{Logical indicating whether decision threshold is a maximum (\code{TRUE}) or minimum (\code{FALSE})}
}
\value{
Dataframe of time and duration of first violation of the threshold
}
\description{
Calculate time of threshold exceedance a single set of simulations
}
