% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fan.R
\name{caption_time}
\alias{caption_time}
\title{Generate time caption for time-outcome fan plot}
\usage{
caption_time(time_outcomes, Dt_max)
}
\arguments{
\item{time_outcomes}{Dataframe or list of dataframe(s) generated by calculate_time()}

\item{Dt_max}{Logical indicating whether decision threshold is a maximum (\code{TRUE}) or minimum (\code{FALSE})}
}
\value{
A character vector or list of character vector(s) of the time caption
}
\description{
Generate time caption for time-outcome fan plot
}
