% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fan.R
\name{caption_time_1}
\alias{caption_time_1}
\title{Generate time caption for time-outcome fan plot for a single set of simulations}
\usage{
caption_time_1(time_outcomes, Dt_max)
}
\arguments{
\item{time_outcomes}{Dataframe or list of dataframe(s) generated by calculate_time()}

\item{Dt_max}{Logical indicating whether decision threshold is a maximum (\code{TRUE}) or minimum (\code{FALSE})}
}
\value{
A character vector of the time caption
}
\description{
Generate time caption for time-outcome fan plot for a single set of simulations
}
