% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plot_descr.R
\name{gen_stand_descr}
\alias{gen_stand_descr}
\title{Generate standard description for all plots}
\usage{
gen_stand_descr(plot, link = "link")
}
\arguments{
\item{plot}{A plot}

\item{link}{A link to the bookdown}
}
\value{
A plot with the standard description caption added
}
\description{
Generate standard description for all plots
}
