% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fan.R
\name{plot_fan}
\alias{plot_fan}
\title{Generate time-outcome fan plot}
\usage{
plot_fan(psa_data, tmin, tmax, Dt, Dt_max)
}
\arguments{
\item{psa_data}{A data.frame (or a list of data.frames containing) where the first column contains the model time, and subsequent columns contain the predicted output for each simulation run at the respective time. See Data format section for more information.}

\item{tmin}{Numeric value giving the minimum simulation time}

\item{tmax}{Numeric value giving the maximum simulation time}

\item{Dt}{Numeric vector of decision thresholds at each time between \code{tmin} and `tmax``}

\item{Dt_max}{Logical indicating whether decision threshold is a maximum (\code{TRUE}) or minimum (\code{FALSE})}
}
\value{
ggplot or list of ggplot(s) time-outcome fan plot(s)
}
\description{
Generate time-outcome fan plot
}
\section{Data format}{


The model time in the first column must contain numeric values indicating a simulation time (ex. 0, 1, 2, 3,...) or dates (ex. 01/01/2024, 01/02/2024) which must be in \code{R} Date format (i.e., class="Date"). If comparing simulation results across multiple scenarios, provide a list of data.frames, where each data.frame gives the output for one scenario.
}

