% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fan.R
\name{plot_fan_1}
\alias{plot_fan_1}
\title{Plot time-outcome fan plot for a single set of simulations}
\usage{
plot_fan_1(psa_data, tmin, tmax, Dt, Dt_max)
}
\arguments{
\item{psa_data}{A data.frame (or a list of data.frames containing) where the first column contains the model time, and subsequent columns contain the predicted output for each simulation run at the respective time. See Data format section for more information.}

\item{tmin}{Numeric value giving the minimum simulation time}

\item{tmax}{Numeric value giving the maximum simulation time}

\item{Dt}{Numeric vector of decision thresholds at each time between \code{tmin} and `tmax``}

\item{Dt_max}{Logical indicating whether decision threshold is a maximum (\code{TRUE}) or minimum (\code{FALSE})}
}
\value{
ggplot time-outcome fan plot
}
\description{
Plot time-outcome fan plot for a single set of simulations
}
