% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_risk.R
\name{tabulate_risk}
\alias{tabulate_risk}
\title{Output risk measures to table}
\usage{
tabulate_risk(exp_risk_list, n_s)
}
\arguments{
\item{exp_risk_list}{List of risk values for each scenario, whether the first element is the risk for the baseline scenario}

\item{n_s}{Numeric value of the number of scenarios (including the baseline) to include in the table}
}
\value{
A character matrix of risk values and policy risk impact (\%) for each strategy
}
\description{
Output risk measures to table
}
